# public/apidocjs/swagger.json
{"swagger":"2.0","info":{"title":"Scheduler Inventory API Docs","version":"1.0.0","description":"API for managing product inventory and warehouses"},"paths":{"/productTypes/{id}":{"get":{"tags":["Product_Types"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Product type ID"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/productTypes":{"get":{"tags":["Product_Types"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"page","in":"path","required":false,"type":"number","description":"Page number for pagination"},{"name":"limit","in":"path","required":false,"type":"number","description":"Number of product types per page"},{"name":"search","in":"path","required":false,"type":"string","description":"Search term for product types"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetProductTypes"}}}}}},"/products":{"post":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/CreateProduct"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}},"get":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"page","in":"path","required":false,"type":"number","description":"Page number for pagination"},{"name":"limit","in":"path","required":false,"type":"number","description":"Number of products per page"},{"name":"search","in":"path","required":false,"type":"string","description":"Search term for products"},{"name":"product_type_id","in":"path","required":false,"type":"string","description":"Filter by product type ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetProducts"}}}}}},"/product/{id}/delete":{"patch":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Product ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/DeleteProduct"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/products/{id}":{"get":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Product ID"},{"name":"token","in":"path","required":true,"type":"string","description":"User token"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/product/{id}/restore":{"patch":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Product ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/RestoreProduct"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/product":{"put":{"tags":["Products"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateProduct"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Product"}}}}}},"/stocks":{"post":{"tags":["Stocks"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/CreateStock"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Stock"}}}}},"get":{"tags":["Stocks"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"product_id","in":"path","required":true,"type":"string","description":"Product ID (optional)"},{"name":"warehouse_id","in":"path","required":true,"type":"string","description":"Warehouse ID (optional)"},{"name":"quantity","in":"path","required":true,"type":"number","description":"Stock quantity (optional)"},{"name":"active","in":"path","required":true,"type":"string","description":"Stock active status (optional)"},{"name":"page","in":"path","required":true,"type":"number","description":"Page number for pagination (optional)"},{"name":"limit","in":"path","required":true,"type":"number","description":"Number of stocks per page (optional)"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetStocks"}}}}},"put":{"tags":["Stocks"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateStock"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Stock"}}}}}},"/stocks/delete":{"patch":{"tags":["Stocks"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/DeleteStock"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Stock"}}}}}},"/stocks/restore":{"patch":{"tags":["Stocks"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/RestoreStock"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Stock"}}}}}},"/warehouses/{id}":{"get":{"tags":["Warehouses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Warehouse ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/Warehouse"}}}}}},"/warehouses":{"get":{"tags":["Warehouses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"page","in":"path","required":false,"type":"number","description":"Page number"},{"name":"limit","in":"path","required":false,"type":"number","description":"Number of warehouses per page"},{"name":"search","in":"path","required":false,"type":"string","description":"Search string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetWarehouses"}}}}}}},"definitions":{"GetProductType":{"properties":{"id":{"type":"string","description":"Product type ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"Product":{"properties":{},"required":[]},"GetProductTypes":{"properties":{"page":{"type":"number","description":"Page number for pagination"},"limit":{"type":"number","description":"Number of product types per page"},"search":{"type":"string","description":"Search term for product types"},"Array":{"type":"array","description":"of product type entries","items":{"type":"Object"}}},"required":["Array"]},"CreateProduct":{"properties":{"code":{"type":"string","description":"Product code"},"description":{"type":"string","description":"Product description"},"product_type_id":{"type":"string","description":"Product type ID"},"image":{"type":"file","description":"Product image"},"token":{"type":"string","description":"User token"}},"required":["code","description","product_type_id","image","token"]},"GetProducts":{"properties":{"page":{"type":"number","description":"Page number for pagination"},"limit":{"type":"number","description":"Number of products per page"},"search":{"type":"string","description":"Search term for products"},"product_type_id":{"type":"string","description":"Filter by product type ID"},"Array":{"type":"array","description":"of product entries","items":{"type":"Object"}}},"required":["Array"]},"DeleteProduct":{"properties":{"id":{"type":"string","description":"Product ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"GetProduct":{"properties":{"id":{"type":"string","description":"Product ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"RestoreProduct":{"properties":{"id":{"type":"string","description":"Product ID"},"token":{"type":"string","description":"User token"}},"required":["id","token"]},"UpdateProduct":{"properties":{"id":{"type":"string","description":"Product ID"},"name":{"type":"string","description":"Product name"},"product_type_id":{"type":"string","description":"Product type ID"},"image_name":{"type":"file","description":"Product image"},"token":{"type":"string","description":"User token"}},"required":["id","name","product_type_id","image_name","token"]},"CreateStock":{"properties":{"product_id":{"type":"string","description":"Product ID"},"warehouse_id":{"type":"string","description":"Warehouse ID"},"quantity":{"type":"number","description":"Stock quantity"},"token":{"type":"string","description":"User token"}},"required":["product_id","warehouse_id","quantity","token"]},"Stock":{"properties":{},"required":[]},"GetStocks":{"properties":{"product_id":{"type":"string","description":"Product ID (optional)"},"warehouse_id":{"type":"string","description":"Warehouse ID (optional)"},"quantity":{"type":"number","description":"Stock quantity (optional)"},"active":{"type":"string","description":"Stock active status (optional)"},"page":{"type":"number","description":"Page number for pagination (optional)"},"limit":{"type":"number","description":"Number of stocks per page (optional)"},"Array":{"type":"array","description":"of stock entries","items":{"type":"Object"}}},"required":["product_id","warehouse_id","quantity","active","page","limit","Array"]},"UpdateStock":{"properties":{"product_id":{"type":"string","description":"Product ID"},"warehouse_id":{"type":"string","description":"Warehouse ID"},"quantity":{"type":"number","description":"Stock quantity"},"token":{"type":"string","description":"User token"}},"required":["product_id","warehouse_id","quantity","token"]},"DeleteStock":{"properties":{"product_id":{"type":"string","description":"Product ID"},"warehouse_id":{"type":"string","description":"Warehouse ID"},"token":{"type":"string","description":"User token"}},"required":["product_id","warehouse_id","token"]},"RestoreStock":{"properties":{"product_id":{"type":"string","description":"Product ID"},"warehouse_id":{"type":"string","description":"Warehouse ID"},"token":{"type":"string","description":"User token"}},"required":["product_id","warehouse_id","token"]},"GetWarehouse":{"properties":{"id":{"type":"string","description":"Warehouse ID"}},"required":["id"]},"Warehouse":{"properties":{},"required":[]},"GetWarehouses":{"properties":{"page":{"type":"number","description":"Page number"},"limit":{"type":"number","description":"Number of warehouses per page"},"search":{"type":"string","description":"Search string"},"Array":{"type":"array","description":"of warehouse entries","items":{"type":"Object"}}},"required":["Array"]}}}